解决spring security跨域问题，与spring boot普通不一样

+++ bigoo-admin/src/main/java/com/jokey/bigoo/admin/entity/RoleMenu.java
@@ -20,12 +20,12 @@ public class RoleMenu implements Serializable {
      * 角色ID
      */
     @ApiModelProperty("角色ID")
-    private BigInteger roleId;
+    private Long roleId;
 
     /**
      * 菜单/按钮ID
      */
     @ApiModelProperty("菜单/按钮ID")
-    private BigInteger menuId;
+    private Long menuId;
 
 }
