+++ bigoo-admin/src/main/java/com/jokey/bigoo/admin/entity/UserRole.java
@@ -20,12 +20,12 @@ public class UserRole implements Serializable {
      * 用户ID
      */
     @ApiModelProperty("用户ID")
-    private BigInteger userId;
+    private Long userId;
 
     /**
      * 角色ID
      */
     @ApiModelProperty("角色ID")
-    private BigInteger roleId;
+    private Long roleId;
 
 }
